refactor: exception 클래스 하나의 패키지로 이동
