+++ src/main/java/lotto/domain/exception/InvalidLottoNumberSizeException.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.exception;
 
 public class InvalidLottoNumberSizeException extends RuntimeException {
