+++ src/main/java/lotto/domain/exception/NotNumberStringException.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.exception;
 
 public class NotNumberStringException extends RuntimeException {
