+++ src/main/java/lotto/domain/exception/NotZeroOrMoreNumberException.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.exception;
 
 public class NotZeroOrMoreNumberException extends RuntimeException {
