+++ src/main/java/lotto/domain/exception/UnpurchasableAmountOfMoneyException.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.exception;
 
 public class UnpurchasableAmountOfMoneyException extends RuntimeException {
