+++ src/test/java/lotto/domain/LottoTest.java
@@ -1,5 +1,7 @@
 package lotto.domain;
 
+import lotto.domain.exception.InvalidLottoNumberSizeException;
+import lotto.domain.exception.OutOfRangeLottoNumberException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
